requirements.txt
dash-table==5.0.0
dash-html-components==2.0.0
dash-core-components==2.0.0
zipp==3.20.2
urllib3==2.2.3
typing-extensions==4.12.2
tenacity==9.0.0
six==1.16.0
setuptools==75.3.0
packaging==24.1
nest-asyncio==1.6.0
MarkupSafe==3.0.2
itsdangerous==2.2.0
idna==3.10
colorama==0.4.6
charset-normalizer==3.4.0
certifi==2024.8.30
blinker==1.8.2
Werkzeug==3.0.6
retrying==1.3.4
requests==2.32.3
plotly==5.24.1
Jinja2==3.1.4
importlib-metadata==8.5.0
click==8.1.7
Flask==3.0.3
dash==2.18.2
redis==5.2.0
PyMuPDF==1.24.13
fire==0.7.0
fonttools==4.54.1
lxml==5.3.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
pdf2docx==0.5.8
python-docx==1.1.2
termcolor==2.5.0
annotated-types==0.7.0
pydantic==2.9.2
pydantic-core==2.23.4
pydantic_settings==2.6.1
python-dotenv==1.0.1